requirements/requirements.txt
click==8.0.1
pyyaml==6.0.1
requests==2.31.0
wheel>=0.22
jsonschema==3.0.2
gitpython==3.1.32
